[hw3] Change labels to long tensor

+++ hw3/ft.py
@@ -307,7 +307,7 @@ def tokenize_gpt2_batch(tokenizer, x, y):
 
         labels.append(sequence_labels.tolist())
 
-    tokenized_sequences["labels"] = torch.IntTensor(labels)
+    tokenized_sequences["labels"] = torch.LongTensor(labels)
 
     return tokenized_sequences
